test(ui-date-time-input): asdsad3
Closes: INSTUI-4014
